fix: changed the `ContentView`
