--- Content Generator/ContentView.swift
@@ -1,26 +1,0 @@
-//
-//  ContentView.swift
-//  Content Generator
-//
-//  Created by Maykon Meneghel on 12/06/23.
-//
-
-import SwiftUI
-
-struct ContentView: View {
-    var body: some View {
-        VStack {
-            Image(systemName: "globe")
-                .imageScale(.large)
-                .foregroundColor(.accentColor)
-            Text("Hello, world!")
-        }
-        .padding()
-    }
-}
-
-struct ContentView_Previews: PreviewProvider {
-    static var previews: some View {
-        ContentView()
-    }
-}
